프로젝트 업로드

+++ src/main/java/com/fontservice/fontory/repository/QuoteRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface QuoteRepository extends JpaRepository<Quote, Integer> {
+public interface QuoteRepository extends JpaRepository<Quote, Long> {
 
     // 가장 최근 문구 한 개 가져오기
     Quote findTopByOrderByCreatedAtDesc();
